+++ hikka/inline/list.py
@@ -2,6 +2,7 @@ import asyncio
 import functools
 import logging
 import time
+import traceback
 from types import FunctionType
 from typing import List as _List
 from typing import Optional, Union
@@ -15,9 +16,11 @@ from aiogram.types import (
     InputTextMessageContent,
 )
 from aiogram.utils.exceptions import RetryAfter
-from telethon.tl.types import Message
 
-from .. import utils
+from telethon.tl.types import Message
+from telethon.errors.rpcerrorlist import ChatSendInlineForbiddenError
+
+from .. import utils, main
 from .types import InlineMessage, InlineUnit
 
 logger = logging.getLogger(__name__)
@@ -108,16 +111,16 @@ class List(InlineUnit):
             ttl = self._markup_ttl
             logger.debug("Defaulted ttl, because it breaks out of limits")
 
-        unit_uid = utils.rand(16)
+        unit_id = utils.rand(16)
         btn_call_data = {key: utils.rand(10) for key in {"back", "next", "close"}}
 
         perms_map = self._find_caller_sec_map() if not manual_security else None
 
-        self._units[unit_uid] = {
+        self._units[unit_id] = {
             "type": "list",
             "chat": None,
             "message_id": None,
-            "uid": unit_uid,
+            "uid": unit_id,
             "btn_call_data": btn_call_data,
             "current_index": 0,
             "strings": strings,
@@ -133,8 +136,8 @@ class List(InlineUnit):
 
         default_map = {
             **(
-                {"ttl": self._units[unit_uid]["ttl"]}
-                if "ttl" in self._units[unit_uid]
+                {"ttl": self._units[unit_id]["ttl"]}
+                if "ttl" in self._units[unit_id]
                 else {}
             ),
             **({"always_allow": always_allow} if always_allow else {}),
@@ -149,7 +152,7 @@ class List(InlineUnit):
                 functools.partial(
                     self._list_back,
                     btn_call_data=btn_call_data,
-                    unit_uid=unit_uid,
+                    unit_id=unit_id,
                 )
             ),
             **default_map,
@@ -159,7 +162,7 @@ class List(InlineUnit):
             "handler": asyncio.coroutine(
                 functools.partial(
                     self._delete_unit_message,
-                    unit_uid=unit_uid,
+                    unit_id=unit_id,
                 )
             ),
             **default_map,
@@ -170,7 +173,7 @@ class List(InlineUnit):
                 functools.partial(
                     self._list_next,
                     btn_call_data=btn_call_data,
-                    unit_uid=unit_uid,
+                    unit_id=unit_id,
                 )
             ),
             **default_map,
@@ -186,24 +189,47 @@ class List(InlineUnit):
         else:
             status_message = None
 
+        async def answer(msg: str):
+            nonlocal message
+            if isinstance(message, Message):
+                await (message.edit if message.out else message.respond)(msg)
+            else:
+                await self._client.send_message(message, msg)
+
         try:
-            q = await self._client.inline_query(self.bot_username, unit_uid)
+            q = await self._client.inline_query(self.bot_username, unit_id)
             m = await q[0].click(
                 utils.get_chat_id(message) if isinstance(message, Message) else message,
                 reply_to=message.reply_to_msg_id
                 if isinstance(message, Message)
                 else None,
             )
-        except Exception:
-            logger.exception("Error sending inline list")
-            del self._units[unit_uid]
-            return
+        except ChatSendInlineForbiddenError:
+            await answer("🚫 <b>You can't send inline units in this chat</b>")
+        except Exception as e:
+            logger.exception("Can't send list")
 
-        await self._units[unit_uid]["future"].wait()
-        del self._units[unit_uid]["future"]
+            if not self._db.get(main.__name__, "inlinelogs", True):
+                msg = f"<b>🚫 List invoke failed! More info in logs</b>"
+            else:
+                exc = traceback.format_exc()
+                # Remove `Traceback (most recent call last):`
+                exc = "\n".join(exc.splitlines()[1:])
+                msg = (
+                    f"<b>🚫 List invoke failed!</b>\n\n"
+                    f"<b>🧾 Logs:</b>\n<code>{exc}</code>"
+                )
 
-        self._units[unit_uid]["chat"] = utils.get_chat_id(m)
-        self._units[unit_uid]["message_id"] = m.id
+            del self._units[unit_id]
+            await answer(msg)
+
+            return False
+
+        await self._units[unit_id]["future"].wait()
+        del self._units[unit_id]["future"]
+
+        self._units[unit_id]["chat"] = utils.get_chat_id(m)
+        self._units[unit_id]["message_id"] = m.id
 
         if isinstance(message, Message) and message.out:
             await message.delete()
@@ -211,27 +237,27 @@ class List(InlineUnit):
         if status_message and not message.out:
             await status_message.delete()
 
-        return InlineMessage(self, unit_uid, self._units[unit_uid]["inline_message_id"])
+        return InlineMessage(self, unit_id, self._units[unit_id]["inline_message_id"])
 
     async def _list_back(
         self,
         call: CallbackQuery,
         btn_call_data: _List[str] = None,
-        unit_uid: str = None,
+        unit_id: str = None,
     ):
-        if not self._units[unit_uid]["current_index"]:
+        if not self._units[unit_id]["current_index"]:
             await call.answer("No way back", show_alert=True)
             return
 
-        self._units[unit_uid]["current_index"] -= 1
+        self._units[unit_id]["current_index"] -= 1
 
         try:
             await self.bot.edit_message_text(
                 inline_message_id=call.inline_message_id,
-                text=self._units[unit_uid]["strings"][
-                    self._units[unit_uid]["current_index"]
+                text=self._units[unit_id]["strings"][
+                    self._units[unit_id]["current_index"]
                 ],
-                reply_markup=self._list_markup(unit_uid),
+                reply_markup=self._list_markup(unit_id),
             )
             await call.answer()
         except RetryAfter as e:
@@ -249,24 +275,24 @@ class List(InlineUnit):
         call: CallbackQuery,
         btn_call_data: _List[str] = None,
         func: FunctionType = None,
-        unit_uid: str = None,
+        unit_id: str = None,
     ):
-        self._units[unit_uid]["current_index"] += 1
+        self._units[unit_id]["current_index"] += 1
         # If we exceeded limit in list
-        if self._units[unit_uid]["current_index"] >= len(
-            self._units[unit_uid]["strings"]
+        if self._units[unit_id]["current_index"] >= len(
+            self._units[unit_id]["strings"]
         ):
             await call.answer("No entries left...", show_alert=True)
-            self._units[unit_uid]["current_index"] -= 1
+            self._units[unit_id]["current_index"] -= 1
             return
 
         try:
             await self.bot.edit_message_text(
                 inline_message_id=call.inline_message_id,
-                text=self._units[unit_uid]["strings"][
-                    self._units[unit_uid]["current_index"]
+                text=self._units[unit_id]["strings"][
+                    self._units[unit_id]["current_index"]
                 ],
-                reply_markup=self._list_markup(unit_uid),
+                reply_markup=self._list_markup(unit_id),
             )
             await call.answer()
         except RetryAfter as e:
@@ -280,33 +306,33 @@ class List(InlineUnit):
             await call.answer("Error occurred", show_alert=True)
             return
 
-    def _list_markup(self, unit_uid: str) -> InlineKeyboardMarkup:
+    def _list_markup(self, unit_id: str) -> InlineKeyboardMarkup:
         """Converts `btn_call_data` into a aiogram markup"""
         markup = InlineKeyboardMarkup()
         markup.add(
             *(
                 [
                     InlineKeyboardButton(
-                        f"⏪ [{self._units[unit_uid]['current_index']} / {len(self._units[unit_uid]['strings'])}]",
-                        callback_data=self._units[unit_uid]["btn_call_data"]["back"],
+                        f"⏪ [{self._units[unit_id]['current_index']} / {len(self._units[unit_id]['strings'])}]",
+                        callback_data=self._units[unit_id]["btn_call_data"]["back"],
                     )
                 ]
-                if self._units[unit_uid]["current_index"] > 0
+                if self._units[unit_id]["current_index"] > 0
                 else []
             ),
             InlineKeyboardButton(
                 "❌",
-                callback_data=self._units[unit_uid]["btn_call_data"]["close"],
+                callback_data=self._units[unit_id]["btn_call_data"]["close"],
             ),
             *(
                 [
                     InlineKeyboardButton(
-                        f"⏩ [{self._units[unit_uid]['current_index'] + 2} / {len(self._units[unit_uid]['strings'])}]",
-                        callback_data=self._units[unit_uid]["btn_call_data"]["next"],
+                        f"⏩ [{self._units[unit_id]['current_index'] + 2} / {len(self._units[unit_id]['strings'])}]",
+                        callback_data=self._units[unit_id]["btn_call_data"]["next"],
                     ),
                 ]
-                if self._units[unit_uid]["current_index"]
-                < len(self._units[unit_uid]["strings"]) - 1
+                if self._units[unit_id]["current_index"]
+                < len(self._units[unit_id]["strings"]) - 1
                 else []
             ),
         )